+++ ChangeLog.md
@@ -6,6 +6,7 @@
 - for --language check that primary font has coverage
 - check normalized lang is in fontconfig orth list
 - new pango --sample-text option with --language
+- add --list-langs option to list all fc orths
 
 ## 0.2.1 (2026-03-26)
 - check and warn about characters without font coverage